+++ ladim/grid.py
@@ -8,6 +8,12 @@ class Grid(Module):
     def ingrid(self, X, Y):
         raise NotImplementedError
 
+    def sample_metric(self, X, Y):
+        raise NotImplementedError
+
+    def atsea(self, X, Y):
+        raise NotImplementedError
+
 
 class RomsGrid(Grid):
     def __init__(self, model: Model, **conf):